RF sensor physical Texas Instruments: implementation
We remove the old `XBee_CC2530_Sensor_Physical` implementation
completely.
