+++ control_panel/Control_Panel_Loading_View.py
@@ -59,7 +59,7 @@ class Control_Panel_Loading_View(Control_Panel_View):
                 except KeyError:
                     # Not an XBee device, so it must be a CC2531 device.
                     self._controller.usb_manager.get_cc2531_device()
-                    rf_sensor_class = "XBee_CC2530_Sensor_Physical"
+                    rf_sensor_class = "RF_Sensor_Physical_Texas_Instruments"
 
             # Reload the sensor class.
             settings = self._controller.arguments.get_settings("control_panel")
