+++ zigbee/RF_Sensor_Physical.py
@@ -36,3 +36,15 @@ class RF_Sensor_Physical(RF_Sensor):
 
         self._ntp = NTP(self)
         self._ntp_delay = self._settings.get("ntp_delay")
+
+    def discover(self, callback):
+        """
+        Discover all RF sensors in the network. The `callback` function is
+        called when an RF sensor reports its identity.
+
+        Classes that inherit this base class must extend this method.
+        """
+
+        super(RF_Sensor_Physical, self).discover(callback)
+
+        self._discovery_callback = callback
