+++ zigbee/XBee_Sensor.py
@@ -9,7 +9,8 @@ from TDMA_Scheduler import TDMA_Scheduler
 
 # pylint: disable=undefined-all-variable
 __all__ = [
-    "XBee_CC2530_Sensor_Physical", "XBee_Sensor_Physical", "RF_Sensor_Simulator"
+    "RF_Sensor_Physical_Texas_Instruments", "XBee_Sensor_Physical",
+    "RF_Sensor_Simulator"
 ]
 
 class SensorClosedError(Exception):
